chore: Update Redis connection URL, cacheMiddleware.js, and productController.js
This commit updates the Redis connection URL in the `redis.js` file to use the IP address `127.0.0.1` instead of the hostname `localhost`. It also includes changes to the `cacheMiddleware.js` file to fix a missing newline at the end of the file and updates the `productController.js` file to remove a console.log statement. These changes improve the reliability and maintainability of the codebase.